trusight_github_ready/requirements.txt
faiss-cpu
fastapi
uvicorn
python-dotenv
pydantic
gunicorn
requests
beautifulsoup4
streamlit
pandas
sentence-transformers
